% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuessR Package.R
\name{reg.uptake}
\alias{reg.uptake}
\title{Calculate a regional CO2 uptake constant from empirical DIC data}
\usage{
reg.uptake(year1, year2, d13c.change)
}
\arguments{
\item{year1}{The year in which samples were collected for the first set of DIC d13C observations.}

\item{year2}{The year in which samples were collected for the second set of DIC d13C observations.}

\item{d13c.change}{The observed change in DIC d13C between the two years.}
}
\value{
Figure out what should go in the 'VALUES' section for these functions.
}
\description{
Calculate a regional CO2 uptake constant from empirical DIC data
}
\details{
Figure out what should go in the 'DETAILS' section for these functions.
}
\examples{
year1 <- 1970
year2 <- 1980
d13c.change <- 0.3

reg.uptake(year1 = year1, year2 = year2, d13c.change = d13c.change)
}
