% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuessR Package.R
\name{SuessR.custom}
\alias{SuessR.custom}
\title{Calculate Suess and Laws corrections for data from a custom region}
\usage{
SuessR.custom(data, custom.region.data, correct.to = 1850)
}
\arguments{
\item{data}{A dataframe including sample ID, year of sample collection, uncorrected d13C
data, and region. This function is specifically for data from regions not
currently built into SuessR. Columns must be named 'id', 'year', 'd13c',
and 'region'.}

\item{custom.region.data}{A dataframe containing environmental data for the custom region from which the
samples originated. Must contain columns titled 'year', 'region', 'esub1', 'esubneg1', 'esub2', 'r', 'C',
'P', "mu', 'beta', 'sst', 'S', and 'CO2atm'. See manual for detailed information on how to supply these
parameters appropriately.}

\item{correct.to}{The year to which the d13C data will be corrected. Defaults to 1850, pre-Suess effect.}
}
\value{
Figure out what should go in the 'VALUES' section for these functions.
}
\description{
The SuessR.custom() function generates region-specific Suess, Laws, and net (Suess + Laws) corrections for
d13C data input by the user. The net correction is then used to calculate the corrected d13C data, which are supplied
in the output. This function is specifically for data from regions not currently built into SuessR
(i.e., different from "Bering", "Aleutians", and "Gulf of Alaska" as of February 2020).
}
\details{
Figure out what should go in the 'DETAILS' section for these functions.
}
\examples{
example.region.data <- data.frame(year = seq(from = 1850, to = 2019, by = 1),
                                                         region = rep("Example Region", 170),
                                                         esub1 = rep(1, 170),
                                                         esubneg1 = rep(1, 170),
                                                         esub2 = rep(26.5, 170),
                                                         r = rep(5, 170),
                                                         C = rep(1.09e-09, 170),
                                                         P = rep(4.57e-10, 170),
                                                         mu = rep(0.5, 170),
                                                         beta = rep(0.2, 170),
                                                         sst = seq(5.9, 6.6, 170),
                                                         S = seq(32.3, 32.7, 170),
                                                         CO2atm = SuessR.reference.data$CO2atm[1:170])

example.custom.data <- data.frame(id = c("Sample 1", "Sample 2", "Sample 3", "Sample 4"),
                                                          year = c(1850, 1900, 1950, 2000),
                                                          d13c = c(-12, -12, -12, -12),
                                                          region = rep("Example Region", 4)

SuessR.custom(data = example.custom.data, custom.region.data = example.region.data, correct.to = 1850)
}
