% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuessR Package.R
\name{SuessR}
\alias{SuessR}
\title{Calculate Suess and Laws corrections for data from a built-in region}
\usage{
SuessR(data, correct.to = 1850)
}
\arguments{
\item{data}{A dataset including sample ID ('id'), year of sample collection ('year'), uncorrected d13C data ('d13C'),
and region ('region').}

\item{correct.to}{The year to which the d13C data will be corrected. Defaults to 1850, pre-Suess effect.}
}
\value{
The output of this function is a data frame that includes the sample ID ('id'), year ('year'),
uncorrected d13C ('d13c.uncor'), Laws correction ('Laws.cor'), Suess Correction ('Suess.cor'), net correction
('net.cor'), and corrected d13C ('d13c.cor') for each sample.
}
\description{
The SuessR() function generates region-specific Suess, Laws, and net (Suess + Laws) corrections for
d13C data input by the user. The net correction is then used to calculate the corrected d13C data, which are supplied
in the output. This function is specifically for data from regions currently built into SuessR
("Bering", "Aleutians", and "Gulf of Alaska" as of February 2020).
}
\details{
Figure out what should go in the 'DETAILS' section for these functions.
}
\examples{
example.data <- data.frame(id = c("Sample 1", "Sample 2", "Sample 3", "Sample 4", "Sample 5", "Sample 6"),
                           year = c(2017, 2017, 2017, 1977, 1977, 1977),
                           d13c = c(-12, -12, -12, -12, -12, -12),
                           region = c("Bering", "Aleutians", "Gulf of Alaska", "Bering", "Aleutians", "Gulf of Alaska"))

SuessR(data = example.data)
}
