% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuessR Reference Data.R
\docType{data}
\name{SuessR.reference.data}
\alias{SuessR.reference.data}
\title{Reference dataset for the SuessR Package}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{

}
\usage{
data(SuessR.reference.data)
}
\description{
Reference dataset for the SuessR Package
}
\examples{
data(SuessR.reference.data)
sst.bering <- SuessR.reference.data$sst[SuessR.reference.data$region == "Bering"]

}
\references{

}
\keyword{datasets}
